Adding more generic configurations to example. Fixed issue with environment variable race conditions with vunit multi threading with a complete hack.
When using multiple threads in vunit, there was a potential for a race condition. If thread #1 set it's environment, but was interrupted by thread #2 before the vunit simulator could copy the environment then thread #1 would run with thread #2's environment variables. The hack to address this was to create a Process for running the simulator The process would set it's environment variables then invoke the simulator.